back to circularqueue

+++ src/circular_queue.h
@@ -4,10 +4,10 @@
 #include <cassert>
 #include "ints.h"
 
-template<typename T, u16 capacity>
+template<typename T, u32 capacity>
 class CircularQueue{
     T* data;
-    std::atomic<u16> head, tail;
+    std::atomic<u32> head, tail;
 public:
     CircularQueue() : head(0), tail(0){
         assert( ( (capacity-1) & capacity ) == 0);
@@ -22,7 +22,7 @@ public:
     inline bool full(){
         return ((head + 1) & (capacity - 1)) == tail;
     }
-    inline u16 count(){
+    inline u32 count(){
         if(head < tail){
             return (capacity - tail) + head;
         }
@@ -33,7 +33,7 @@ public:
         head = ((head + 1) & (capacity - 1));
     }
     inline void set_push(const T& item){
-        for(u16 i = tail; i != head; i = (i + 1) & (capacity - 1)){
+        for(u32 i = tail; i != head; i = (i + 1) & (capacity - 1)){
             if(data[i] == item){
                 return;
             }
